+++ bindings/python/src/processors.rs
@@ -159,7 +159,7 @@ pub struct PyBertProcessing {}
 #[pymethods]
 impl PyBertProcessing {
     #[new]
-    fn new(sep: (String, u32), cls: (String, u32)) -> (Self, PyPostProcessor) {
+    fn new(sep: (String, u64), cls: (String, u64)) -> (Self, PyPostProcessor) {
         (
             PyBertProcessing {},
             PyPostProcessor::new(Arc::new(BertProcessing::new(sep, cls).into())),
@@ -203,8 +203,8 @@ impl PyRobertaProcessing {
     #[new]
     #[pyo3(signature = (sep, cls, trim_offsets = true, add_prefix_space = true))]
     fn new(
-        sep: (String, u32),
-        cls: (String, u32),
+        sep: (String, u64),
+        cls: (String, u64),
         trim_offsets: bool,
         add_prefix_space: bool,
     ) -> (Self, PyPostProcessor) {
@@ -262,9 +262,9 @@ impl From<PySpecialToken> for SpecialToken {
 
 impl FromPyObject<'_> for PySpecialToken {
     fn extract(ob: &PyAny) -> PyResult<Self> {
-        if let Ok(v) = ob.extract::<(String, u32)>() {
+        if let Ok(v) = ob.extract::<(String, u64)>() {
             Ok(Self(v.into()))
-        } else if let Ok(v) = ob.extract::<(u32, String)>() {
+        } else if let Ok(v) = ob.extract::<(u64, String)>() {
             Ok(Self(v.into()))
         } else if let Ok(d) = ob.downcast::<PyDict>() {
             let id = d
@@ -274,7 +274,7 @@ impl FromPyObject<'_> for PySpecialToken {
             let ids = d
                 .get_item("ids")
                 .ok_or_else(|| exceptions::PyValueError::new_err("`ids` must be specified"))?
-                .extract::<Vec<u32>>()?;
+                .extract::<Vec<u64>>()?;
             let tokens = d
                 .get_item("tokens")
                 .ok_or_else(|| exceptions::PyValueError::new_err("`tokens` must be specified"))?
